Remove inclusion of index_string.hpp from table.hpp
This will prevent re-compilation of all test files when updating the
interface to index, which should be internal only.
